rename dscore to nryant_dscore to avoid name collision (#97)
